[CCUBE-1546][MAHI]Remove test cases and added breakpoint tokens to grid styling
